# dashboard-v2/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Honeypot dashboard-v2 BFF
  version: 0.1.0-source-audit
  description: >-
    Source-backed description of the dashboard-v2 same-origin BFF. See API.md
    for the complete field and semantic contract. GET routes are read-only;
    auth methods only establish or expire a dashboard cookie.
servers:
  - url: /
    description: Same-origin dashboard deployment
...
components:
  securitySchemes:
    DashboardSession:
      type: apiKey
      in: cookie
      name: dashboard_v2_session
      description: HMAC-backed HttpOnly dashboard session cookie.
  parameters:
...
    SessionId:
      name: session_id
      in: query
      required: true
      schema: {type: string, maxLength: 256}
...
  schemas:
    Error:
      type: object
      properties:
        error: {type: string}
        route: {type: string}
      additionalProperties: true
...
    JsonRecord:
      type: object
      additionalProperties: true
...
    AiAdvisoryResponse:
      type: object
      properties:
        ok: {type: boolean}
        status: {type: string}
        session_id: {type: string}
        advisory: {$ref: '#/components/schemas/JsonRecord'}
        metrics: {$ref: '#/components/schemas/JsonRecord'}
        timestamp: {type: string, format: date-time}
      additionalProperties: true
...
  responses:
    Unauthorized:
      description: Dashboard session missing or invalid.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    NotFound:
      description: Route or requested data was not found.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    BadGateway:
      description: Upstream response was not safe JSON or exceeded 8 MiB.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unavailable:
      description: Upstream unavailable or origin not configured safely.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
paths:
...
  /api/ai-advisory:
    get:
      tags: [monitor]
      security: [{DashboardSession: []}]
      summary: Stored AI advisory detail
      x-upstream-path: /api/ai-advisory
      parameters: [{$ref: '#/components/parameters/SessionId'}]
      responses: {200: {description: Advisory detail, content: {application/json: {schema: {$ref: '#/components/schemas/AiAdvisoryResponse'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '502': {$ref: '#/components/responses/BadGateway'}, '503': {$ref: '#/components/responses/Unavailable'}}
...